+++ src/components/setting/organization/organization-list.vue
@@ -523,30 +523,31 @@ export default {
     },
     deleteById: function (relationId, type) {
       let that = this
-      this.$dialog('询问', '您确定删除该记录吗？', true, true,
-        () => { // 点击确定
-          Vue.$ajax({
-            method: 'delete',
-            url: Vue.$adminServerURL + '/OrganizationController/delete/' + that.selectedNode.id + '/' + relationId + '/' + type
-          }).then(res => {
-            if (res.data.flag !== 'SUCCESS') {
-              this.$message({
-                message: res.data.message,
-                type: 'warning'
-              })
-            } else {
-              this.$message({
-                message: res.data.message,
-                type: 'success'
-              })
-              this.listPage()
-            }
-          })
-        },
-        () => { // 点击关闭
 
-        }
-      )
+      this.$confirm('您确定删除该记录吗', '询问', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        Vue.$ajax({
+          method: 'delete',
+          url: Vue.$adminServerURL + '/OrganizationController/delete/' + that.selectedNode.id + '/' + relationId + '/' + type
+        }).then(res => {
+          if (res.data.flag !== 'SUCCESS') {
+            this.$message({
+              message: res.data.message,
+              type: 'warning'
+            })
+          } else {
+            this.$message({
+              message: res.data.message,
+              type: 'success'
+            })
+            this.listPage()
+          }
+        })
+      }).catch(() => {
+      })
     },
     setResource: function (data) {
       if (this.showFunctionTreePopover) {
